fix: remove Decimal from swapQuote. refactor codebase
