+++ packages/dynamic-bonding-curve/src/constants.ts
@@ -1,22 +1,21 @@
 import BN from 'bn.js'
 import { PublicKey } from '@solana/web3.js'
 
-export const BASIS_POINT_MAX: number = 10_000
+// Curve + Math constants
 export const OFFSET: number = 64
 export const U64_MAX: BN = new BN('18446744073709551615')
 export const MIN_SQRT_PRICE: BN = new BN('4295048016')
 export const MAX_SQRT_PRICE: BN = new BN('79226673521066979257578248091')
 
-// Curve constants
-export const MAX_CURVE_POINT = 20
-export const FEE_DENOMINATOR = new BN(1_000_000_000)
-export const MAX_FEE_NUMERATOR = new BN(500_000_000)
-export const MAX_TOKEN_SUPPLY = new BN(10_000_000_000)
-export const SCALE_OFFSET = 64
-export const ONE = new BN(1).shln(SCALE_OFFSET)
 export const RESOLUTION = 64
 export const ONE_Q64 = new BN(1).shln(RESOLUTION)
-export const PARTNER_SURPLUS_SHARE = 90
+export const FEE_DENOMINATOR = 1_000_000_000
+export const MAX_FEE_NUMERATOR = 500_000_000 // 50%
+export const BASIS_POINT_MAX = 10000
+export const MAX_CURVE_POINT = 16
+export const PARTNER_SURPLUS_SHARE = 80 // 80%
+export const SWAP_BUFFER_PERCENTAGE = 25 // 25%
+export const MAX_SWALLOW_PERCENTAGE = 20 // 20%
 
 // Pubkey
 export const DYNAMIC_BONDING_CURVE_PROGRAM_ID = new PublicKey(
