+++ packages/dynamic-bonding-curve/src/utils.ts
@@ -23,16 +23,6 @@ import {
 import { TokenType } from './types'
 import BN from 'bn.js'
 
-export function getSecondKey(key1: PublicKey, key2: PublicKey) {
-    const buf1 = key1.toBuffer()
-    const buf2 = key2.toBuffer()
-    // Buf1 > buf2
-    if (Buffer.compare(buf1, buf2) === 1) {
-        return buf2
-    }
-    return buf1
-}
-
 export function getFirstKey(key1: PublicKey, key2: PublicKey) {
     const buf1 = key1.toBuffer()
     const buf2 = key2.toBuffer()
@@ -41,6 +31,16 @@ export function getFirstKey(key1: PublicKey, key2: PublicKey) {
         return buf1
     }
     return buf2
+}
+
+export function getSecondKey(key1: PublicKey, key2: PublicKey) {
+    const buf1 = key1.toBuffer()
+    const buf2 = key2.toBuffer()
+    // Buf1 > buf2
+    if (Buffer.compare(buf1, buf2) === 1) {
+        return buf2
+    }
+    return buf1
 }
 
 /**
